requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=1.0
flask-cors>=3.0
